+++ src/cylinder.rs
@@ -1,22 +1,25 @@
 //! A cylinder structure.
 
 use substrate::Sheet;
-use system::{Coord, Component, IntoComponent, Residue, Translate};
+use system::{Coord, Component, IntoComponent, ResidueBase, Translate};
 
-/// A cylinder of some `Residue`s.
-pub struct Cylinder<'a> {
+#[derive(Clone, Debug)]
+/// A cylinder of some residues.
+pub struct Cylinder {
     /// `Cylinder` origin, positioned in the middle point of one of the cylinder edges.
-    /// `Residue` positions are relative to this.
+    /// Residue positions are relative to this.
     pub origin: Coord,
     /// Cylinder radius.
     pub radius: f64,
     /// Cylinder height.
     pub height: f64,
-    /// `Residue`s belonging to the `Cylinder`.
-    pub residues: Vec<Residue<'a>>,
+    /// Residue base.
+    pub residue_base: ResidueBase,
+    /// Residue positions of the `Cylinder`.
+    pub residues: Vec<Coord>,
 }
 
-impl<'a> Cylinder<'a> {
+impl Cylinder {
     /// Fold a `Sheet` into a `Cylinder`.
     ///
     /// The `Sheet` is folded along the x axis using its `size` as the full length.
@@ -28,49 +31,59 @@ impl<'a> Cylinder<'a> {
     /// of positions along z is discarded to create the cylinder. It would be preferable
     /// for this to be accounted for in some way, but it would require knowledge of the
     /// number of layers.
-    pub fn from_sheet(sheet: &'a Sheet) -> Cylinder<'a> {
+    pub fn from_sheet(sheet: &Sheet) -> Cylinder {
         let (length, height, _) = sheet.size.to_tuple();
 
         let radius = length / (2.0 * ::std::f64::consts::PI);
         let origin = sheet.origin;
 
-        let residues = sheet.residues
+        let residues = sheet.residue_coords
             .iter()
             .map(|res| {
-                let (x, y, _) = res.position.to_tuple();
+                let (x, y, _) = res.to_tuple();
                 let angle = (x * 360.0 / length).to_radians();
 
                 let x = radius * angle.sin();
                 let z = -radius * angle.cos();
 
-                Residue {
-                    base: res.base,
-                    position: Coord::new(x, y, z),
-                }
+                Coord::new(x, y, z)
             }).collect();
 
-        Cylinder { origin, radius, height, residues }
+        Cylinder { origin, radius, height, residue_base: sheet.residue_base.clone(), residues }
     }
 }
 
-impl<'a> IntoComponent<'a> for Cylinder<'a> {
-    fn into_component(self) -> Component<'a> {
+impl IntoComponent for Cylinder {
+    fn to_component(&self) -> Component {
         let radius = self.radius;
         let height = self.height;
 
         Component {
             origin: self.origin,
             box_size: Coord::new(2.0 * radius, height, 2.0 * radius),
-            residues: self.residues,
+            residue_base: self.residue_base.clone(),
+            residue_coords: self.residues.clone(),
+        }
+    }
+
+    fn into_component(self) -> Component {
+        let radius = self.radius;
+        let height = self.height;
+
+        Component {
+            origin: self.origin,
+            box_size: Coord::new(2.0 * radius, height, 2.0 * radius),
+            residue_base: self.residue_base,
+            residue_coords: self.residues,
         }
     }
 
     fn num_atoms(&self) -> usize {
-        self.residues.iter().map(|r| r.base.atoms.len()).sum()
+        self.residue_base.atoms.len() * self.residues.len()
     }
 }
 
-impl<'a> Translate for Cylinder<'a> {
+impl Translate for Cylinder {
     fn translate(mut self, trans: &Coord) -> Self {
         self.origin = self.origin + *trans;
         self
@@ -93,18 +106,19 @@ mod tests {
         }
     }
 
-    fn setup_sheet<'a>(residue: &'a ResidueBase) -> Sheet<'a> {
+    fn setup_sheet(residue: &ResidueBase) -> Sheet {
         // A sheet with four positions along x will result in a cylinder in the x-z plane
         // with points at 0, 90, 180 and 270 degrees and its center at the first position.
         // Positions along y will be unchanged.
         Sheet {
             origin: Coord::new(0.0, 0.0, 0.0),
             size: Coord::new(4.0, 1.0, 0.0),
-            residues: vec![
-                Residue { base: &residue, position: Coord::new(0.0, 0.0, 0.0) },
-                Residue { base: &residue, position: Coord::new(1.0, 0.5, 0.0) },
-                Residue { base: &residue, position: Coord::new(2.0, 1.0, 0.0) },
-                Residue { base: &residue, position: Coord::new(3.0, 1.5, 0.0) },
+            residue_base: residue.clone(),
+            residue_coords: vec![
+                Coord::new(0.0, 0.0, 0.0),
+                Coord::new(1.0, 0.5, 0.0),
+                Coord::new(2.0, 1.0, 0.0),
+                Coord::new(3.0, 1.5, 0.0),
             ]
         }
     }
@@ -122,10 +136,10 @@ mod tests {
         assert_eq!(radius, cylinder.radius);
         assert_eq!(1.0, cylinder.height); // The size along y
 
-        assert_eq!(Coord::new(0.0, 0.0, -radius), cylinder.residues[0].position);
-        assert_eq!(Coord::new(radius, 0.5, 0.0), cylinder.residues[1].position);
-        assert_eq!(Coord::new(0.0, 1.0, radius), cylinder.residues[2].position);
-        assert_eq!(Coord::new(-radius, 1.5, 0.0), cylinder.residues[3].position);
+        assert_eq!(Coord::new(0.0, 0.0, -radius), cylinder.residues[0]);
+        assert_eq!(Coord::new(radius, 0.5, 0.0), cylinder.residues[1]);
+        assert_eq!(Coord::new(0.0, 1.0, radius), cylinder.residues[2]);
+        assert_eq!(Coord::new(-radius, 1.5, 0.0), cylinder.residues[3]);
     }
 
     #[test]
@@ -138,18 +152,20 @@ mod tests {
         assert_eq!(cylinder.origin, sheet.origin);
 
         let radius = 4.0 / (2.0 * PI);
-        assert_eq!(Coord::new(0.0, 0.0, -radius), cylinder.residues[0].position);
-        assert_eq!(Coord::new(radius, 0.5, 0.0), cylinder.residues[1].position);
-        assert_eq!(Coord::new(0.0, 1.0, radius), cylinder.residues[2].position);
-        assert_eq!(Coord::new(-radius, 1.5, 0.0), cylinder.residues[3].position);
+        assert_eq!(Coord::new(0.0, 0.0, -radius), cylinder.residues[0]);
+        assert_eq!(Coord::new(radius, 0.5, 0.0), cylinder.residues[1]);
+        assert_eq!(Coord::new(0.0, 1.0, radius), cylinder.residues[2]);
+        assert_eq!(Coord::new(-radius, 1.5, 0.0), cylinder.residues[3]);
     }
 
     #[test]
     fn cylinder_from_empty_sheet_works() {
+        let residue = setup_residue();
         let sheet = Sheet {
             origin: Coord::new(0.0, 0.0, 0.0),
             size: Coord::new(1.0, 2.0, 3.0),
-            residues: vec![],
+            residue_base: residue,
+            residue_coords: vec![],
         };
 
         let cylinder = Cylinder::from_sheet(&sheet);